separe les fonctions de dataset dans un fichier à part pour alleger training.c
